renamed ProjectLoadingActions.js to ProjectDataLoadingActions.js
